Added option to cache passages.  uses the transients API, uses "esv" . md5(URL that fetches the passage). If container or class is set to empty the element or attribute is not added.
Added option expire_seconds to specify the number of seconds to cache the passages.  Its value can be suffixed with "m", "h", "d", or "w" for minute, hour, day, or week, respectively.


Added option limit, if nonzero sets limit of size of the passage to cache.  Passages larger than this limit will not be cached.

Added option debug, if true adds a message indicating whether the entry came from the cache or the server.

Added remove option.