Fix problem with where next Sunday is in next year.
In make_passage_timestamp replaced %w/%W code using make_time with an offset in seconds from the current time to the target day.  I think a Julian date of > 366 was causing a problem.  The result was that the date resolved to today's date during the week before 1/1/2017, which is a Sunday.  Not thuroughly tested.