Added 'key' attribute.
In esv() and esv_ref() there is now an attribute 'key' which defaults to the key stored in the options page.

The previous rev was supposed to have removed "test" as the default key, which it only partially did.  This rev removes the rest, so if you want to remove the "key" option you will have to deal with use of "test" as a default.

Changed version to 1.1.3.